Added new inference files
