--- docker-retrieval-agent/retrieval-table/requirements.txt
@@ -1,7 +1,0 @@
-fastapi
-uvicorn
-chromadb
-numpy
-scipy
-scikit-learn
-sentence-transformers
